2 things added:
1. Added static values to UserFormFragment for user name field and user
photo field.

-If a user added a photo, OnActivityResult method call resulted in
leaving the MainActivity, which called FragmentChange to detect if user
had valid info. FragmentChange was being called and would wipe what data
the user had already had in the name field and in the ImageView.

-Fixed by having static values that would save the name text and the
image bitmap when onActivityResult is called, and changing UI elements
to these static values in OnViewCreated

2. A pair of list views to display the emails of users in the pending
invite list and partner list.

-Hoping to change to custom layout of name and picture with buttons to
accept or decline pending invites