1. Improved overall stability
- Created a DoneVariable class to detect when a new user has been
created. When both a new user and a photo has successfully been
uploaded, the fragment is changed.
-Before, a datachange listener was executing to change the fragment.
This would cause crashes because a change would be detected and a
fragment change would occur even when operations were not yet done
in the user creation fragment.
-Arrays to populate the listviews are now checked if null to
prevent null exceptions when setting the list adapters.

2. Made some minor UI changes, a profile picture is required when
creating a user.